requirements.txt
djangorestframework
djangorestframework-simplejwt
redis
celery
drf-yasg
Pillow
psycopg2-binary
pytelegrambotapi
python-decouple
